DAOS-5479 dmg/vmd: Set FAULT state on VMD LED when device is evicted
Add transport address param to the dmg storage set nvme-faulty command.
$dmg storage set nvme-faulty --uuid=DEV_UUID --traddr=TRADDR

When a device is manually evicted via the command above, also set the
LED device state to a FAULTY ("on") state if the device is a VMD
supported device. If the device is just a regular NVMe deivce, then
the device will just be evicted.

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>